# api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: CLM-Core
  version: 1.0.0
  description: API endpoints the clm-core module offers
servers:
  - url: '{scheme}://{hostname}:{port}{path}'
    description: The production API server
    variables:
      hostname:
        default: localhost
      port:
        default: '5008'
      path:
        default: ''
      scheme:
        default: http
security:
  - bearerAuth: []
paths:
  /core/authentication:
    post:
      tags:
        - pblc
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  default: fame@fokus.fraunhofer.de
                password:
                  type: string
                  default: 12345
              required:
                - email
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    $ref: '#/components/parameters/accessToken'
                  refreshToken:
                    $ref: '#/components/parameters/accessToken'
                  accessTokenExpiresIn:
                    type: string
                    default: 12/10/2022
                  refreshTokenExpiresIn:
                    type: string
                    default: 12/10/2022
...
  /core/mgmt/roles:
    get:
      tags:
        - mgmt-roles
        - mgmt
      description: Get all roles
      summary: Get all roles in the system
      responses:
        '200':
          description: All roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/role'
    post:
      tags:
        - mgmt-roles
        - mgmt
      description: Add a role to the system
      summary: 'Add a role to the system [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      requestBody:
        description: The user to add to the group
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createUpdateRole'
      responses:
        '200':
          description: Successfully added user to group
...
  /core/mgmt/consumers:
    get:
      tags:
        - mgmt-consumers
        - mgmt
      description: Super-Admin can get all API-Tokens
      summary: 'Get all API-Tokens [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: All API-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/consumer'
    post:
      tags:
        - mgmt-consumers
        - mgmt
      description: Super-Admin can get create API-Tokens
      summary: 'Create an API-Token [Minimum Role: SUPER-ADMIN]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully created API-Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/consumer'
...
  /core/mgmt/groups:
    get:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can access all his groups and subgroups
      summary: 'Get all groups [Minimum Role: LEARNER]'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully removed user from group
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/group'
    post:
      tags:
        - mgmt-groups
        - mgmt
      description: A user can create a group if hes atleast ADMIN in another group
      summary: 'Create group [Minimum Role: ADMIN]'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                displayName:
                  type: string
                  description: how the group should be named
                  default: Example-Group
                role:
                  type: string
                  enum:
                    - LEARNER
                    - INSTRUCTOR
                    - ADMIN
                  description: The role you want to give the group
                  default: LEARNER
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully removed user from group
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/group'
...
  /core/mgmt/users:
    get:
      tags:
        - mgmt-users
        - mgmt
      description: User can get all the users he is in the same group with
      summary: 'Get users [Minimum Role: ''LEARNER'']'
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: The list of users
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/user'
    post:
      tags:
        - mgmt-users
        - mgmt
      description: User can create a user when he is in an admin group
      summary: 'Create a user [Minimum Role: "ADMIN"]'
      requestBody:
        required: true
        description: Payload user
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  default: student@localhost.de
                password:
                  type: string
                  default: '12345'
                givenName:
                  type: string
                  default: Maximillian
                familyName:
                  type: string
                  default: Mustermann
      parameters:
        - $ref: '#/components/parameters/accessToken'
      responses:
        '200':
          description: Successfully created a user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
...
  /core/users:
    post:
      tags:
        - pblc
      description: >-
        Register user. The instructions will send to the email provided in the
        payload
      summary: >-
        Register user. The instructions will send to the email provided in the
        payload
      requestBody:
        content:
          application/json:
            schema:
              properties:
                givenName:
                  type: string
                  default: Max
                familyName:
                  type: string
                  default: Mustermann
                password:
                  type: string
                  default: 12345
                email:
                  type: string
                  default: Max@Mustermann.de
      responses:
        '200':
          description: >-
            Successfully registerd user (not yet activated). Instructions will
            be sent to the provided E-Mail
components:
  schemas:
...
    role:
      type: object
      properties:
        _id:
          type: string
          description: id of the group
          default: jaiosdjfiojifj2341dij
        displayName:
          type: string
          description: How to role is named
          default: Learner
        updatedAt:
          type: string
          description: when the group was last updated
          default: 10-10-2020
        createdAt:
          type: string
          description: id of the group
          default: 10-10-2022
        resourcePermissions:
          type: object
          description: The permissions of the group
          default:
            lo: 15
            service: 15
            tool: 15
            group: 15
            user: 15
            role: 15
            consumer: 15
        lineage:
          type: boolean
          description: If the group is a lineage group
          default: true
        strength:
          type: number
          description: The strength of the group
          default: 0
    createUpdateRole:
      type: object
      properties:
        lineage:
          type: boolean
          description: If the group is a lineage group
          default: true
        displayName:
          type: string
          default: Learner
          description: How to role is named
        strength:
          type: number
          description: The strength of the group
          default: 0
        resourcePermissions:
          type: object
          description: The permissions of the group
          properties:
            lo:
              type: number
              description: The permission for the learning object
              default: 1
            service:
              type: number
              description: The permission for the service
              default: 1
            tool:
              type: number
              description: The permission for the tool
              default: 1
            group:
              type: number
              description: The permission for the group
              default: 1
            user:
              type: number
              description: The permission for the user
              default: 1
            role:
              type: number
              description: The permission for the role
              default: 1
            consumer:
              type: number
              description: The permission for the consumer
              default: 1
    consumer:
      properties:
        displayName:
          type: string
          default: API-Token
        active:
          type: boolean
          default: false
        userId:
          type: boolean
          description: The user who associates with that API-Token
          default: consumer@localhost.org
        domain:
          type: string
          default: STAKEHOLDER-DOMAIN
        paths:
          type: object
          properties:
            scope:
              type: array
              items:
                type: string
                enum:
                  - GET
                  - PUT
                  - POST
                  - DELETE
                  - PATCH
              description: The HTTP methods which can be applied to the route
            route:
              type: string
              default: /some/route/the/token/can/access
    group:
      type: object
      properties:
        _id:
          type: string
          description: id of the group
          default: jaiosdjfiojifj2341dij
        displayName:
          type: string
          description: how the group should be displayed
          default: Example-Group
        updatedAt:
          type: string
          description: when the group was last updated
          default: 10-10-2020
        createdAt:
          type: string
          description: id of the group
          default: 10-10-2022
    user:
      type: object
      properties:
        email:
          type: string
          default: fame@fokus.fraunhofer.de
        isVerified:
          type: boolean
          default: true
        givenName:
          type: string
          default: max
        familyName:
          type: string
          default: Mustermann
        isSuperAdmin:
          type: boolean
          default: false
  parameters:
    accessToken:
      name: x-access-token
      in: header
      description: The access token
      required: true
      example: exampleAccessToken
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshAuth:
      type: apiKey
      in: header
      name: x-refresh-token
...